+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.5.3 - 2021-5-16
+### Added
+- Add `AbstractBlockStateMixin` which allows for blocks with the "bookshelves"
+  tag to affect enchanting tables.
+- Add `PlumeTags` which makes it simple to access platform specific tags.
+
 ## v0.5.2 - 2021-05-15
 ### Added
 - Add `PlumeLadderBlock` which makes vanilla `LadderBlock` public.